+++ flight/PiOS/posix/pios_flyingpio.c
@@ -54,10 +54,10 @@ const struct pios_rcvr_driver pios_flyingpio_rcvr_driver = {
         .read = PIOS_FLYINGPIO_Receiver_Get,
 };
 
-static bool PIOS_FLYINGPIO_ADC_Available(uint32_t dev_int, uint32_t pin);
-static int32_t PIOS_FLYINGPIO_ADC_PinGet(uint32_t dev_int, uint32_t pin);
-static uint8_t PIOS_FLYINGPIO_ADC_NumberOfChannels(uint32_t dev_int);
-static float PIOS_FLYINGPIO_ADC_LSB_Voltage(uint32_t dev_int);
+static bool PIOS_FLYINGPIO_ADC_Available(uintptr_t dev_int, uint32_t pin);
+static int32_t PIOS_FLYINGPIO_ADC_PinGet(uintptr_t dev_int, uint32_t pin);
+static uint8_t PIOS_FLYINGPIO_ADC_NumberOfChannels(uintptr_t dev_int);
+static float PIOS_FLYINGPIO_ADC_LSB_Voltage(uintptr_t dev_int);
 
 const struct pios_adc_driver pios_flyingpio_adc_driver = {
 		.available = PIOS_FLYINGPIO_ADC_Available,
@@ -377,7 +377,7 @@ int32_t PIOS_FLYINGPIO_Receiver_Get(uintptr_t dev_int, uint8_t channel)
 	return dev->rcvr_value[channel];
 }
 
-static bool PIOS_FLYINGPIO_ADC_Available(uint32_t dev_int, uint32_t pin) {
+static bool PIOS_FLYINGPIO_ADC_Available(uintptr_t dev_int, uint32_t pin) {
 	if (pin >= FPPROTO_MAX_ADCCHANS) {
 		return false;
 	}
@@ -385,7 +385,7 @@ static bool PIOS_FLYINGPIO_ADC_Available(uint32_t dev_int, uint32_t pin) {
 	return true;
 }
 
-static int32_t PIOS_FLYINGPIO_ADC_PinGet(uint32_t dev_int, uint32_t pin) {
+static int32_t PIOS_FLYINGPIO_ADC_PinGet(uintptr_t dev_int, uint32_t pin) {
 	pios_flyingpio_dev_t dev = (pios_flyingpio_dev_t) dev_int;
 
 	if (pin >= FPPROTO_MAX_ADCCHANS) {
@@ -395,12 +395,12 @@ static int32_t PIOS_FLYINGPIO_ADC_PinGet(uint32_t dev_int, uint32_t pin) {
 	return dev->adc_value[pin];
 }
 
-static uint8_t PIOS_FLYINGPIO_ADC_NumberOfChannels(uint32_t dev_int) {
+static uint8_t PIOS_FLYINGPIO_ADC_NumberOfChannels(uintptr_t dev_int) {
 	return FPPROTO_MAX_ADCCHANS;
 }
 
 #define VREF_PLUS 3.3f
-static float PIOS_FLYINGPIO_ADC_LSB_Voltage(uint32_t dev_int) {
+static float PIOS_FLYINGPIO_ADC_LSB_Voltage(uintptr_t dev_int) {
 	// Value is expected to be a fraction of 3V3
         return VREF_PLUS / (((uint32_t)1 << 16) - 1);
 }
