+++ flight/targets/revolution/fw/pios_board.c
@@ -619,7 +619,7 @@ void PIOS_Board_Init(void) {
 #endif    /* PIOS_INCLUDE_I2C */
 
 #if defined(PIOS_INCLUDE_ADC)
-	uint32_t internal_adc_id;
+	uintptr_t internal_adc_id;
 	PIOS_INTERNAL_ADC_Init(&internal_adc_id, &pios_adc_cfg);
 	PIOS_ADC_Init(&pios_internal_adc_id, &pios_internal_adc_driver, internal_adc_id);
